clean up ununsed variables and renamed screens
